Move types `TokenBundleSize{Assessor,Assessment}` to `Tx.Constraints`.
These types relate to constraints on the sizes of token bundles in the
context of transaction outputs.

So it makes sense to locate them together with other functions and types
that relate to transaction size constraints.
