+++ lib/core/cardano-wallet-core.cabal
@@ -276,7 +276,6 @@ library
       Cardano.Wallet.Primitive.Types.Tx.CBOR.Hash
       Cardano.Wallet.Primitive.Types.Tx.Constraints
       Cardano.Wallet.Primitive.Types.Tx.SealedTx
-      Cardano.Wallet.Primitive.Types.Tx.TokenBundleAssessments
       Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
       Cardano.Wallet.Primitive.Types.Tx.Tx
       Cardano.Wallet.Primitive.Types.Tx.TxMeta
