+++ lib/shelley/test/unit/Cardano/Wallet/Shelley/CompatibilitySpec.hs
@@ -71,10 +71,10 @@ import Cardano.Wallet.Primitive.Types.TokenBundle.Gen
     ( genTokenBundle, genTokenBundleSmallRange, shrinkTokenBundleSmallRange )
 import Cardano.Wallet.Primitive.Types.Tx
     ( TxSize (..), txOutMaxCoin, txOutMinCoin )
+import Cardano.Wallet.Primitive.Types.Tx.Constraints
+    ( TokenBundleSizeAssessment (..), TokenBundleSizeAssessor (..) )
 import Cardano.Wallet.Primitive.Types.Tx.Gen
     ( genTxOutTokenBundle )
-import Cardano.Wallet.Primitive.Types.Tx.TokenBundleAssessments
-    ( TokenBundleSizeAssessment (..), TokenBundleSizeAssessor (..) )
 import Cardano.Wallet.Shelley.Compatibility
     ( CardanoBlock
     , StandardCrypto
